requirements.txt
# Core scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
tabula-py==2.9.0
camelot-py==0.10.1
openpyxl==3.1.2
pypdf==3.17.1

# New dependencies for modernized system
pyyaml==6.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional dependencies for enhanced features
aiohttp==3.9.1
requests-cache==1.1.1
python-multipart==0.0.6
jinja2==3.1.2

# Development dependencies
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
